Update main.cpp
Added support for specifying a resolution and refresh rate in the config file. Added checks for a failed result from CreateWindow that skip showing and destroying the window. The resolution capability isn't quite as powerful as gamestream launchpad though that power comes at a cost (launchpad checks every 2 seconds to see if the window it launched is active and if so confirms the resolution hasn't been changed and then changes it, which is functionality I really think should lie in the launcher, specifically PlayNite which I intend to add in a patch eventually.) For well behaved games that work in borderless mode this resolution and refresh rate setting capability should be sufficient.