Update main.cpp
added a config option to just always choose the highest refresh rate. Changed the method of resolution change to set it in the registry (this eliminates some strangeness with some misbehaving games (Assasin's Creed Valhalla is one,) that ignore that use the registry to reset after running instead of the current values when they start up. and don't seem to function correctly in borderless mode.) By default the previous registry settings are restored after execution but there is an option to disable this. The reasoning behind this is for people like me who set unsupported moonlight resolutions like 1440p. Gamestream sets the resolution to 720p so I don't want to return to that resolution after exit. I also now check to see if I'm trying to change to the current resolution and skip the resolution change in that case.

With these changes I think I've scratched all my personal itches and have no need for gamestream_launchpad anymore. I took a look at Playnite and it has a display resolution changer extension I'll likely modify it to always choose the highest refresh rate (would anyone playing games ever not want to highest supported refresh rate?)